feat!: resolve MSE show names to IDs using the directory
